Move Decaf code to decaf.rs
