+++ src/lib.rs
@@ -44,8 +44,9 @@ extern crate rand;
 // Modules for low-level operations directly on field elements and curve points.
 
 pub mod field;
-pub mod curve;
 pub mod scalar;
+pub mod curve;
+pub mod decaf;
 
 // Constant-time functions and other miscelaneous utilities.
